refacto(pallet_starknet): split tests between multiple files
