+++ crates/pallets/starknet/src/tests/declare_tx.rs
@@ -1,0 +1,48 @@
+use blockifier::test_utils::{get_contract_class, ERC20_CONTRACT_PATH};
+use frame_support::{assert_err, assert_ok, bounded_vec};
+use hex::FromHex;
+use mp_starknet::execution::{CallEntryPointWrapper, ContractClassWrapper, EntryPointTypeWrapper};
+use mp_starknet::transaction::types::Transaction;
+
+use super::mock::*;
+use crate::Error;
+
+#[test]
+fn given_contract_declare_tx_works_once_not_twice() {
+    new_test_ext().execute_with(|| {
+        System::set_block_number(0);
+        run_to_block(2);
+
+        let none_origin = RuntimeOrigin::none();
+        let (account_addr, _, _) = account_helper(TEST_ACCOUNT_SALT);
+
+        let erc20_class = ContractClassWrapper::from(get_contract_class(ERC20_CONTRACT_PATH));
+        let erc20_class_hash =
+            <[u8; 32]>::from_hex("057eca87f4b19852cfd4551cf4706ababc6251a8781733a0a11cf8e94211da95").unwrap();
+
+        let mut transaction = Transaction {
+            sender_address: account_addr,
+            call_entrypoint: CallEntryPointWrapper::new(
+                Some(erc20_class_hash),
+                EntryPointTypeWrapper::External,
+                None,
+                bounded_vec![],
+                account_addr,
+                account_addr,
+            ),
+            ..Transaction::default()
+        };
+        // Cannot declare a class with None
+        assert_err!(
+            Starknet::declare(none_origin.clone(), transaction.clone()),
+            Error::<Test>::ContractClassMustBeSpecified
+        );
+
+        transaction.contract_class = Some(erc20_class);
+
+        assert_ok!(Starknet::declare(none_origin.clone(), transaction.clone()));
+        // TODO: Uncomment once we have ABI support
+        // assert_eq!(Starknet::contract_class_by_class_hash(erc20_class_hash), erc20_class);
+        assert_err!(Starknet::declare(none_origin, transaction), Error::<Test>::ClassHashAlreadyDeclared);
+    });
+}
