+++ crates/pallets/starknet/src/tests/deploy_account_tx.rs
@@ -1,0 +1,103 @@
+use core::str::FromStr;
+
+use frame_support::{assert_err, assert_ok, bounded_vec, BoundedVec};
+use hex::FromHex;
+use mp_starknet::execution::{CallEntryPointWrapper, EntryPointTypeWrapper};
+use mp_starknet::transaction::types::Transaction;
+use sp_core::{H256, U256};
+
+use super::mock::*;
+use crate::Error;
+
+#[test]
+fn given_contract_run_deploy_account_tx_works() {
+    new_test_ext().execute_with(|| {
+        System::set_block_number(0);
+        run_to_block(2);
+
+        let none_origin = RuntimeOrigin::none();
+        // TEST ACCOUNT CONTRACT
+        // - ref testnet tx(0x0751b4b5b95652ad71b1721845882c3852af17e2ed0c8d93554b5b292abb9810)
+        let salt = "0x03b37cbe4e9eac89d54c5f7cc6329a63a63e8c8db2bf936f981041e086752463";
+        let (test_addr, account_class_hash, calldata) = account_helper(salt);
+
+        let transaction = Transaction {
+            sender_address: test_addr,
+            call_entrypoint: CallEntryPointWrapper::new(
+                Some(account_class_hash),
+                EntryPointTypeWrapper::External,
+                None,
+                BoundedVec::try_from(calldata.clone().into_iter().map(U256::from).collect::<Vec<U256>>()).unwrap(),
+                test_addr,
+                test_addr,
+            ),
+            contract_address_salt: Some(H256::from_str(salt).unwrap()),
+            ..Transaction::default()
+        };
+
+        assert_ok!(Starknet::deploy_account(none_origin, transaction));
+        assert_eq!(Starknet::contract_class_hash_by_address(test_addr).unwrap(), account_class_hash);
+    });
+}
+
+#[test]
+fn given_contract_run_deploy_account_tx_twice_fails() {
+    new_test_ext().execute_with(|| {
+        System::set_block_number(0);
+        run_to_block(2);
+
+        let none_origin = RuntimeOrigin::none();
+        let salt = "0x03b37cbe4e9eac89d54c5f7cc6329a63a63e8c8db2bf936f981041e086752463";
+        let (test_addr, account_class_hash, calldata) = account_helper(salt);
+
+        // TEST ACCOUNT CONTRACT
+        // - ref testnet tx(0x0751b4b5b95652ad71b1721845882c3852af17e2ed0c8d93554b5b292abb9810)
+        let transaction = Transaction {
+            sender_address: test_addr,
+            call_entrypoint: CallEntryPointWrapper::new(
+                Some(account_class_hash),
+                EntryPointTypeWrapper::External,
+                None,
+                BoundedVec::try_from(calldata.clone().into_iter().map(U256::from).collect::<Vec<U256>>()).unwrap(),
+                test_addr,
+                test_addr,
+            ),
+            contract_address_salt: Some(H256::from_str(salt).unwrap()),
+            ..Transaction::default()
+        };
+
+        assert_ok!(Starknet::deploy_account(none_origin.clone(), transaction.clone()));
+        // Check that the account was created
+        assert_eq!(Starknet::contract_class_hash_by_address(test_addr).unwrap(), account_class_hash);
+        assert_err!(Starknet::deploy_account(none_origin, transaction), Error::<Test>::AccountAlreadyDeployed);
+    });
+}
+
+#[test]
+fn given_contract_run_deploy_account_tx_undeclared_then_it_fails() {
+    new_test_ext().execute_with(|| {
+        System::set_block_number(0);
+        run_to_block(2);
+
+        let none_origin = RuntimeOrigin::none();
+        let rand_address =
+            <[u8; 32]>::from_hex("0000000000000000000000000000000000000000000000000000000000001234").unwrap();
+        let undeclared_class_hash =
+            <[u8; 32]>::from_hex("00000000000000000000000000000000000000000000000000000000BEEFDEAD").unwrap();
+
+        let transaction = Transaction {
+            sender_address: rand_address,
+            call_entrypoint: CallEntryPointWrapper::new(
+                Some(undeclared_class_hash),
+                EntryPointTypeWrapper::External,
+                None,
+                bounded_vec![],
+                rand_address,
+                rand_address,
+            ),
+            ..Transaction::default()
+        };
+
+        assert_err!(Starknet::deploy_account(none_origin, transaction), Error::<Test>::TransactionExecutionFailed);
+    });
+}
