+++ crates/pallets/starknet/src/tests/mod.rs
@@ -1,0 +1,9 @@
+mod account_helper;
+mod current_block;
+mod declare_tx;
+mod deploy_account_tx;
+mod erc20;
+mod fee;
+mod invoke_tx;
+
+mod mock;
